pyproject.toml
[tool.poetry]
name = "concourseatom"
version = "0.1.0"
description = "Read Concourse Pipelines and intelligently allowing pipelines to be built from snipits"
authors = ["Ben Greene <BenJGreene@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
"ruamel.yaml" = "^0.17.21"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
Sphinx = "^5.1.1"
sphinx-autodoc-typehints = "^1.19.2"
pytest-sphinx = "^0.5.0"
flake8 = "^5.0.4"
flake8-black = "^0.3.3"
pylint = "^2.15.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
concmerge = "concourseatom.tools:cli"
